The most significant changes involve the update of the `docker-compose.yml` file for the `ThirdPartyFreight.Api` service and the modification of the `ThirdPartyFreight.Api.http` file.
In the `docker-compose.yml` file, several environment variables were added to the service configuration. These include settings for the ASP.NET Core environment, HTTP and HTTPS ports, URLs, and Kestrel certificates. A volume was also added to the service, mapping a local directory to a container directory in read-only mode.

In the `ThirdPartyFreight.Api.http` file, a GET request to the `/weatherforecast` endpoint was removed and replaced with a new GET request to the `/api/v1/agreements` endpoint. This new request includes an `Accept` header set to `application/json`.

Changes:
1. Added several environment variables to the `docker-compose.yml` file for the `ThirdPartyFreight.Api` service. These include `ASPNETCORE_ENVIRONMENT`, `ASPNETCORE_HTTP_PORTS`, `ASPNETCORE_HTTPS_PORTS`, `ASPNETCORE_URLS`, `ASPNETCORE_Kestrel__Certificates__Default__Password`, and `ASPNETCORE_Kestrel__Certificates__Default__Path`.
2. Added a volume to the `docker-compose.yml` file for the `ThirdPartyFreight.Api` service, mapping the local directory `./src/ThirdPartyFreight.Api/PrivateKeys/aspnetapp.pfx` to the container directory `/https/aspnetapp.pfx` in read-only mode.
3. Removed a GET request to the `/weatherforecast` endpoint in the `ThirdPartyFreight.Api.http` file.
4. Added a new GET request to the `/api/v1/agreements` endpoint in the `ThirdPartyFreight.Api.http` file, including an `Accept` header set to `application/json`.